博客查询添加完毕，增加了后台管理时展示的数据username和观看量

+++ src/main/java/com/myblog/blog/entity/Blog.java
@@ -13,8 +13,8 @@ public class Blog {
     private  String content;
     private String firstPicture;
     private String flag;
-    private  Integer views;
-    private  Integer CommentCount;
+    private  long views;
+    private  long CommentCount;
     private boolean appreciation;
     private boolean commentabled;
 
@@ -108,7 +108,7 @@ public class Blog {
         this.flag = flag;
     }
 
-    public Integer getViews() {
+    public long getViews() {
         return views;
     }
 
@@ -116,7 +116,7 @@ public class Blog {
         this.views = views;
     }
 
-    public Integer getCommentCount() {
+    public long getCommentCount() {
         return CommentCount;
     }
